requirements.txt
Django==1.4.3
Flask==0.9
Jinja2==2.6
Pygments==1.6
South==0.7.6
Sphinx==1.1.3
Werkzeug==0.8.3
croniter==0.3.3
django-breadcrumbs==1.1.3-p1
django-redis==3.1.6
django-social-auth==0.7.18
django-tastypie==0.9.14
docutils==0.10
gevent==0.13.8
greenlet==0.4.0
httplib2==0.8
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.4.6
python-dateutil==1.5
python-openid==2.2.5
redis==2.7.2
requests==1.1.0
six==1.3.0
wsgiref==0.1.2
zope.interface==4.0.3
gunicorn==0.17.2
django-heroku-memcacheify==0.3
django-heroku-postgresify==0.2
pylibmc==1.2.3
newrelic==1.11.0.55
django-storages==1.1.7
schedule==0.1.11
rq==0.3.11


